% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_pal.r
\name{cl_pal}
\alias{cl_pal}
\title{cl_pal
:DEPRECATED: now using ch_pal instead}
\usage{
cl_pal(colour)
}
\arguments{
\item{colour}{int colourlovers palette id}
}
\description{
cl_pal
:DEPRECATED: now using ch_pal instead
}
\keyword{DEPRECATED}
