% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_pal.R
\name{image_pal}
\alias{image_pal}
\title{image_pal}
\usage{
image_pal(image_source = "https://www.r-project.org/Rlogo.png")
}
\arguments{
\item{image_source}{character file path or url}

\item{n}{integer number of color centers to extract}
}
\description{
Read an image from disk or url and extract dominant colors using kmeans.
}
