% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_palette.R
\name{make_palette}
\alias{make_palette}
\title{make_palette}
\usage{
make_palette(colour = NULL, n = 7, reverse = FALSE, shuffle = FALSE,
  default = TRUE)
}
\arguments{
\item{colour}{colour(s) to be turned into a palette}

\item{n}{numeric the number of colours to be returned}

\item{reverse}{logical should the return palette order be reversed?}

\item{shuffle}{logical should the return palette be shuffled?}

\item{default}{logical should default palette be Tableau-esque orange-blue diverging?}
}
\value{
list object containing palette and plot
}
\description{
This function will generate colour palettes of arbitrary size
from an R base graphics color function, ColorBrewer named palette, 
COLOURLovers pallete ID or vector of named or hex colours.
Also, a source image file (png or jpeg) or url can be used.
Defaults to a nice Tableau-esque Orange->Blue divering or 
Stephen Few-esque Earth->Emerald...
}
\examples{

### Brewer Colour Palette 
colourgen(colour = 'Spectral')

### COLOURlovers ID
make_palette(colour = 3914747)
make_palette(colour = 987654)

### Custom Colour Vector
my_colours <- c("#CAF60D", "#18D33A", "#4255EC", "#E60873", "#19312A")
make_palette(colour = my_colours)

### Sample Source Image
make_palette(colour = "https://www.r-project.org/Rlogo.png")

### Default Behaviour
make_palette(default = TRUE)
make_palette(default = FALSE)

}
